Move polaris test provider into components
