+++ src/components/PolarisTestProvider/index.ts
@@ -1,0 +1,5 @@
+export {
+  default,
+  Props,
+  WithPolarisTestProviderOptions,
+} from './PolarisTestProvider';
