+++ src/components/PolarisTestProvider/tests/PolarisTestProvider.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
-import {PolarisTestProvider} from '../PolarisTestProvider';
+import {PolarisTestProvider} from '../..';
 
 describe('PolarisTestProvider', () => {
   it("doesn't render in strict mode by default", () => {
