+++ src/test-utilities/react-testing.tsx
@@ -5,7 +5,7 @@ import translations from '../../locales/en.json';
 import {
   PolarisTestProvider,
   WithPolarisTestProviderOptions,
-} from './PolarisTestProvider';
+} from '../components';
 
 export {mount};
